+++ presidio-analyzer/presidio_analyzer/predefined_recognizers/medical_license_recognizer.py
@@ -21,8 +21,9 @@ class MedicalLicenseRecognizer(PatternRecognizer):
     PATTERNS = [
         Pattern(
             "USA DEA Certificate Number (weak)",
-            r"[a-zA-Z]{2}\d{7}|[a-zA-Z]{1}9\d{7}",
-            0.3,
+            r"[abcdefghjklmprstuxABCDEFGHJKLMPRSTUX]{1}[a-zA-Z]{1}\d{7}|"
+            r"[abcdefghjklmprstuxABCDEFGHJKLMPRSTUX]{1}9\d{7}",
+            0.4,
         ),
     ]
 
@@ -61,11 +62,11 @@ class MedicalLicenseRecognizer(PatternRecognizer):
             return [int(dig) for dig in str(n)]
 
         digits = digits_of(sanitized_value[2:])
-        odd_digits = digits[-1::-2]
-        even_digits = digits[-2::-2]
-        checksum = sum(odd_digits)
-        for d in even_digits:
-            checksum += sum(digits_of(str(d * 2)))
+        checksum = digits.pop()
+        even_digits = digits[-1::-2]
+        odd_digits = digits[-2::-2]
+        checksum *= -1
+        checksum += 2 * sum(even_digits) + sum(odd_digits)
         return checksum % 10 == 0
 
     @staticmethod