+++ CHANGELOG.md
@@ -8,6 +8,12 @@ All notable changes to this project will be documented in this file.
 * Added Italian fiscal code recognizer
 * Added Italian driver license recognizer
 
+### Changed
+#### Analyzer
+* Improved MEDICAL_LICENSE pattern and fixed checksum verification
+* Added pattern for next gen US passport in `presidio-analyzer/presidio_analyzer/predefined_recognizers/us_passport_recognizer.py`
+
+
 ## [2.2.29] - 12.07.2022
 ### Added